api/cache: Make cache TTL configurable

+++ packages/api/src/store/cache.ts
@@ -2,20 +2,37 @@ import _ from "lodash";
 import NodeCache from "node-cache";
 
 class Cache {
-  storage = new NodeCache({ stdTTL: 120 });
+  storage: NodeCache;
+
+  init(options?: NodeCache.Options) {
+    if (this.storage) {
+      throw new Error("Cache already initialized");
+    }
+
+    this.storage = new NodeCache({
+      stdTTL: 120,
+      checkperiod: 60,
+      ...options,
+    });
+  }
 
   get<T>(cacheKey: string) {
+    if (!this.storage) return;
+
     const content = this.storage.get(cacheKey) as T;
     // always make copies in case caller mutates the object (yeah we still have that)
     return content && _.cloneDeep(content);
   }
 
   set<T>(cacheKey: string, content: T, ttl?: string | number) {
+    if (!this.storage) return;
+
     content = _.cloneDeep(content);
     this.storage.set(cacheKey, content, ttl);
   }
 
   delete(cacheKey: string) {
+    if (!this.storage) return;
     this.storage.del(cacheKey);
   }
 
@@ -34,6 +51,7 @@ class Cache {
 
   // Test helper to clear the cache
   flush() {
+    if (!this.storage) return;
     this.storage.flushAll();
   }
 }
